COMMERCE-2475 dispatch-talend - .java update, rename classes, refactor packages, use new dispatch entities...
