--- modules/apps/dispatch/dispatch-talend/src/main/java/com/liferay/commerce/data/integration/talend/internal/process/type/TalendProcessType.java
@@ -1,49 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- *
- *
- *
- */
-
-package com.liferay.commerce.data.integration.talend.internal.process.type;
-
-import com.liferay.commerce.data.integration.process.type.ProcessType;
-import com.liferay.portal.kernel.language.LanguageUtil;
-
-import java.util.Locale;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Marco Leo
- */
-@Component(
-	immediate = true,
-	property = {
-		"commerce.data.integration.process.type.key=" + TalendProcessType.KEY,
-		"commerce.data.integration.process.type.order=100"
-	},
-	service = ProcessType.class
-)
-public class TalendProcessType implements ProcessType {
-
-	public static final String KEY = "TALEND";
-
-	@Override
-	public String getKey() {
-		return KEY;
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, getKey());
-	}
-
-}
